requirements.txt
google-genai
openai
pyyaml